Split SettingsPanel.tsx into focused sub-components
Extract AudioSettings, STTSettings, LanguageSettings, TranslatorSettings,
SubtitleSettings, and SessionControls into src/renderer/components/settings/.
Shared types, styles, and utilities moved to shared.ts.
All state remains in the parent SettingsPanel; sub-components receive props.
No behavioral changes — UI and IPC communication preserved exactly.
