+++ src/renderer/components/settings/index.ts
@@ -1,0 +1,7 @@
+export { AudioSettings } from './AudioSettings'
+export { STTSettings } from './STTSettings'
+export { LanguageSettings } from './LanguageSettings'
+export { TranslatorSettings } from './TranslatorSettings'
+export { SubtitleSettings } from './SubtitleSettings'
+export { SessionControls } from './SessionControls'
+export { Section } from './Section'
